+++ src/app/app.component.ts
@@ -15,7 +15,7 @@ export class AppComponent implements OnInit {
   constructor(private http: HttpClient) {}
 
   ngOnInit(): void {
-    this.onChange({target: {value: 'bmisimple'}});
+    this.onChange({target: {value: 'bmicase'}});
   }
 
   onChange(event): void {