+++ CHANGELOG.md
@@ -4,8 +4,8 @@ This project follows [Semantic Versioning](http://semver.org/).
 
 ## [1.4.0] 2021-08-27
 ### Added
-- Copy button and tooltip preview for automatically generated FHIRPath and
-x-fhir-query.
+- Tooltip preview for automatically generated FHIRPath and x-fhir-query.
+- Copy button next to FHIRPath and x-fhir-query preview.
 
 ## [1.3.0] 2021-08-23
 ### Added