+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [1.5.0] 2021-09-09
+### Added
+- Case statement support for the final expression.
+
 ## [1.4.0] 2021-08-27
 ### Added
 - Tooltip preview for automatically generated FHIRPath and x-fhir-query.